bugfix: remove not used code and add source for copied code
