--- pkg/utils/signal_posix.go
@@ -1,11 +1,0 @@
-//go:build !windows
-// +build !windows
-
-package utils
-
-import (
-	"os"
-	"syscall"
-)
-
-var shutdownSignals = []os.Signal{os.Interrupt, syscall.SIGTERM}
