--- pkg/utils/singal.go
@@ -1,39 +1,0 @@
-package utils
-
-import (
-	"context"
-	"os"
-	"os/signal"
-)
-
-var onlyOneSignalHandler = make(chan struct{})
-var shutdownHandler chan os.Signal
-
-// SetupSignalHandler registered for SIGTERM and SIGINT. A stop channel is returned
-// which is closed on one of these signals. If a second signal is caught, the program
-// is terminated with exit code 1.
-// Only one of SetupSignalContext and SetupSignalHandler should be called, and only can
-// be called once.
-func SetupSignalHandler() <-chan struct{} {
-	return SetupSignalContext().Done()
-}
-
-// SetupSignalContext is same as SetupSignalHandler, but a context.Context is returned.
-// Only one of SetupSignalContext and SetupSignalHandler should be called, and only can
-// be called once.
-func SetupSignalContext() context.Context {
-	close(onlyOneSignalHandler) // panics when called twice
-
-	shutdownHandler = make(chan os.Signal, 2)
-
-	ctx, cancel := context.WithCancel(context.Background())
-	signal.Notify(shutdownHandler, shutdownSignals...)
-	go func() {
-		<-shutdownHandler
-		cancel()
-		<-shutdownHandler
-		os.Exit(1) // second signal. Exit directly.
-	}()
-
-	return ctx
-}
